Custom exeptions

+++ HR_Management.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR_Management.Application.Exeptions;
 using HR_Management.Application.Features.LeaveAllocations.Requests.Commands;
 using HR_Management.Application.Persistence.Contract;
+using HR_Management.Domain;
 using MediatR;
 
 namespace HR_Management.Application.Features.LeaveAllocations.Handlers.Commands;
@@ -18,6 +20,10 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+        if (leaveAllocation == null)
+        {
+            throw new NotFoundExeption(nameof(LeaveAllocation), request.Id);
+        }
         await _leaveAllocationRepository.Delete(leaveAllocation);
 
         return Unit.Value;
