+++ HR_Management.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HR_Management.Application.Exeptions;
 using HR_Management.Application.Features.LeaveRequests.Requests.Commands;
 using HR_Management.Application.Persistence.Contract;
+using HR_Management.Domain;
 using MediatR;
 
 namespace HR_Management.Application.Features.LeaveRequests.Handlers.Commands;
@@ -17,6 +19,10 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+        if (leaveRequest == null)
+        {
+            throw new NotFoundExeption(nameof(LeaveRequest), request.Id);
+        }
         await _leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
